requirements.txt
# Core
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-pinecone>=0.2.0
langchain-openai>=0.3.11

# Vector DB
pinecone-client>=6.0.0

# Transformers + HuggingFace
transformers>=4.40.0
accelerate>=0.27.0
sentence-transformers>=2.7.0

# Utilities
tiktoken
flask
GitPython
python-dotenv
PyGithub
streamlit
watchdog
sqlalchemy
